revert change that didn't fix /analyze warning C6297

+++ src/RenderCache.cpp
@@ -124,11 +124,11 @@ static RectI GetTileRectDevice(BaseEngine *engine, int pageNo, int rotation, flo
     RectD mediabox = engine->PageMediabox(pageNo);
 
     if (tile.res && tile.res != INVALID_TILE_RES) {
-        double width = mediabox.dx / (int)(1 << tile.res);
+        double width = mediabox.dx / (1 << tile.res);
         mediabox.x += tile.col * width;
         mediabox.dx = width;
-        double height = mediabox.dy / (int)(1 << tile.res);
-        mediabox.y += ((int)(1 << tile.res) - tile.row - 1) * height;
+        double height = mediabox.dy / (1 << tile.res);
+        mediabox.y += ((1 << tile.res) - tile.row - 1) * height;
         mediabox.dy = height;
     }
